clinet: optimize ssl error log

+++ wss-client-http3.c
@@ -56,7 +56,7 @@ void http_response_cb_v3(struct bufferevent *tev, void *raw) {
         goto error;
     }
     evbuffer_drain(input, frame_length);
-    LOGD("wss is ready for peer %d, remain: %d", get_peer_port(raw), (int) evbuffer_get_length(input));
+    LOGD("wss is ready for peer %d, remain: %zu", get_peer_port(raw), evbuffer_get_length(input));
     tunnel_wss(raw, tev, wss_output_filter_v3);
     return;
 error:
